MOB-233 cleanup
