+++ mobile/lib/screens/phone_authentication/phone_auth_widgets.dart
@@ -12,7 +12,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_svg/svg.dart';
 
-import '../auth/auth_widgets.dart';
+import '../../widgets/auth_widgets.dart';
 import '../email_authentication/email_auth_screen.dart';
 import 'phone_auth_screen.dart';
